+++ WebStore.Domain/Entities/Orders/Order.cs
@@ -1,7 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -21,22 +19,5 @@ namespace WebStore.Domain.Entities.Orders
         public DateTime Date { get; set; } = DateTime.Now;
 
         public IEnumerable<OrderItem> Items { get; set; } = new HashSet<OrderItem>();
-    }
-
-    public class OrderItem : Entity
-    {
-        [Required]
-        public Order Order { get; set; }
-
-        [Required]
-        public Product Product { get; set; }
-
-        [Column(TypeName ="decimal(18,2)")]
-        public decimal Price { get; set; }
-
-        public int Quantity { get; set; }
-
-        [NotMapped]
-        public decimal TotalItmPrice => Price * Quantity;
     }
 }
