+++ WebStore.Domain/Entities/Orders/OrderItem.cs
@@ -1,0 +1,23 @@
+﻿using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
+using WebStore.Domain.Entities.Base;
+
+namespace WebStore.Domain.Entities.Orders
+{
+    public class OrderItem : Entity
+    {
+        [Required]
+        public Order Order { get; set; }
+
+        [Required]
+        public Product Product { get; set; }
+
+        [Column(TypeName ="decimal(18,2)")]
+        public decimal Price { get; set; }
+
+        public int Quantity { get; set; }
+
+        [NotMapped]
+        public decimal TotalItmPrice => Price * Quantity;
+    }
+}
