+++ WebStore/Infrastructure/Services/Interfaces/IOrderService.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using WebStore.Domain.Entities.Orders;
+using WebStore.ViewModels;
+
+namespace WebStore.Infrastructure.Services.Interfaces
+{
+   public interface IOrderService
+    {
+        Task<IEnumerable<Order>> GetUserOrders(string UserName);
+
+        Task<Order> GetOrderById(int id);
+
+        Task<Order> CreateOrder(string UserName, CartViewModel Cart, OrderViewModel OrderModel);
+    }
+}
